refactor tests. place in separate files
